+++ Changelog.md
@@ -13,6 +13,7 @@
 * Add oEmbed content to the mobile view [#4343](https://github.com/diaspora/diaspora/pull/4353)
 * One click to select the invite URL [#4447](https://github.com/diaspora/diaspora/pull/4447)
 * Disable "mark all as read" link if all notifications are read [#4463](https://github.com/diaspora/diaspora/pull/4463)
+* Collapse aspect list and tag followings list when switching to other views [#4462](https://github.com/diaspora/diaspora/pull/4462)
 
 # 0.2.0.0